v3

+++ apps/server/src/models/event.model.ts
@@ -58,11 +58,11 @@ EventModel.init({
     allowNull: false
   },
   longitude: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: false
   },
   latitude: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: false
   },
   category: {
